+++ source/Messaging.IntegrationTests/Transactions/MoveIn/MoveInRequestHandlerTests.cs
@@ -15,34 +15,32 @@
 using System.IO;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using Messaging.Application.Common;
-using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Application.OutgoingMessages;
-using Messaging.Application.Transactions;
 using Messaging.Application.Transactions.MoveIn;
 using Messaging.Application.Xml;
 using Messaging.Application.Xml.SchemaStore;
 using Messaging.IntegrationTests.Fixtures;
+using Messaging.IntegrationTests.IncomingMessages;
 using Messaging.IntegrationTests.TestDoubles;
 using Xunit;
 using Xunit.Categories;
 
-namespace Messaging.IntegrationTests.IncomingMessages
+namespace Messaging.IntegrationTests.Transactions.MoveIn
 {
     [IntegrationTest]
-    public class RequestChangeOfSupplierHandlingTests : TestBase
+    public class MoveInRequestHandlerTests : TestBase
     {
         private readonly IOutgoingMessageStore _outgoingMessageStore;
         private readonly IMoveInTransactionRepository _moveInTransactionRepository;
-        private readonly RequestChangeOfSupplierHandler _requestChangeOfSupplierHandler;
+        private readonly MoveInRequestHandler _moveInRequestHandler;
 
-        public RequestChangeOfSupplierHandlingTests(DatabaseFixture databaseFixture)
+        public MoveInRequestHandlerTests(DatabaseFixture databaseFixture)
             : base(databaseFixture)
         {
             _outgoingMessageStore = GetService<IOutgoingMessageStore>();
             _moveInTransactionRepository =
                 GetService<IMoveInTransactionRepository>();
-            _requestChangeOfSupplierHandler = GetService<RequestChangeOfSupplierHandler>();
+            _moveInRequestHandler = GetService<MoveInRequestHandler>();
         }
 
         [Fact]
@@ -50,7 +48,7 @@ namespace Messaging.IntegrationTests.IncomingMessages
         {
             var incomingMessage = IncomingMessageBuilder.CreateMessage();
 
-            await _requestChangeOfSupplierHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
+            await _moveInRequestHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
 
             var savedTransaction = _moveInTransactionRepository.GetById(incomingMessage.MarketActivityRecord.Id);
             Assert.NotNull(savedTransaction);
@@ -66,7 +64,7 @@ namespace Messaging.IntegrationTests.IncomingMessages
                 .WithConsumerName("John Doe")
                 .Build();
 
-            await _requestChangeOfSupplierHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
+            await _moveInRequestHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
             var confirmMessage = _outgoingMessageStore.GetByOriginalMessageId(incomingMessage.Id)!;
             await RequestMessage(confirmMessage.Id.ToString()).ConfigureAwait(false);
 
@@ -83,7 +81,7 @@ namespace Messaging.IntegrationTests.IncomingMessages
                 .WithConsumerName(null)
                 .Build();
 
-            await _requestChangeOfSupplierHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
+            await _moveInRequestHandler.HandleAsync(incomingMessage).ConfigureAwait(false);
             var rejectMessage = _outgoingMessageStore.GetByOriginalMessageId(incomingMessage.Id)!;
             await RequestMessage(rejectMessage.Id.ToString()).ConfigureAwait(false);
